Refactoring and regoranizing code.
This moves code around so that fetching page titles and formatted twats are
their own functions. This makes it easier for splitting up things like
extracting these things out of messages implicitly as well as having commands
for them.

Doing this reorganization allows the stuff scraped out of messages to be
displayed in the same order they were present in the message. Closes #5, since
it also displays everything scraped on one line as well.