Adding kinder failure to ElasticSearch.forceBulk.
Previously, data could be lost if there was an error, since bulk data gets
discarded as soon as the bulk data insert request is built. Now, failed
requests get logged so they can be inspected and/or retried.

The problem that led to wanting to have this anyway seems to be fixed by
ensuring that a bulk request has more than one operation in it.